+++ packages/components/grid/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './grid';
+
+export { default as version } from './version';
